chore: move backend runai module to api.lib.compute subpackage
